Add bundle files to gitignore
