+++ src/components/extra-films-list.js
@@ -1,0 +1,10 @@
+const createExtraFilmsListMarkup = (title) => {
+  return (
+    `<section class="films-list--extra">
+      <h2 class="films-list__title">${title}</h2>
+      <div class="films-list__container"></div>
+    </section>`
+  );
+};
+
+export {createExtraFilmsListMarkup};
