+++ src/components/show-more-btn.js
@@ -1,0 +1,7 @@
+const createShowMoreBtnMarkup = () => {
+  return (
+    `<button class="films-list__show-more">Show more</button>`
+  );
+};
+
+export {createShowMoreBtnMarkup};
